Refactor commonly used code to helper methods
