+++ app/test/e2e/email-notifications/viirs-fires-alerts-emails.spec.js
@@ -16,6 +16,8 @@ const { createSubscription } = require('../utils/helpers');
 const { createMockViirsAlertsQuery, createMockGeostore } = require('../utils/mock');
 const { ROLES } = require('../utils/test.constants');
 
+const { assertSubscriptionStats } = require('../utils/helpers/email-notifications');
+
 nock.disableNetConnect();
 nock.enableNetConnect(process.env.HOST_IP);
 
@@ -25,20 +27,6 @@ chai.use(require('chai-datetime'));
 const CHANNEL = config.get('apiGateway.queueName');
 const redisClient = redis.createClient({ url: config.get('redis.url') });
 redisClient.subscribe(CHANNEL);
-
-const assertSubscriptionStats = (jsonMessage, sub) => {
-    jsonMessage.should.have.property('sender').and.equal('gfw');
-    jsonMessage.should.have.property('data').and.be.a('object');
-    jsonMessage.data.should.have.property('counter').and.equal(1);
-    jsonMessage.data.should.have.property('dataset').and.equal('viirs-active-fires');
-    jsonMessage.data.should.have.property('users').and.be.an('array').and.length(1);
-    jsonMessage.data.users[0].should.have.property('userId').and.equal(sub.userId);
-    jsonMessage.data.users[0].should.have.property('subscriptionId').and.equal(sub.id);
-    jsonMessage.data.users[0].should.have.property('email').and.equal(sub.resource.content);
-    jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
-    jsonMessage.recipients[0].should.be.an('object').and.have.property('address')
-        .and.have.property('email').and.equal('info@vizzuality.com');
-};
 
 const bootstrapVIIRSAlertTest = () => {
     const beginDate = moment().subtract('1', 'w');
